Refacto Redux store creation
